Added reduce motion configuration option
Allow side panel expand/collapse animations to be disabled Fixes #756

This was the only animation I could find really that wasn't part of an extension or other library. At least we have this option now that we can use to detect this preference. I think it might be useful to store a users preference in local storage or something.