Added helper onAccessibleClick
This helper is a replacement for places where we are using jQuery click or press events. This makes it easy to swap inaccessible instances where you want to be able to navigate with the space bar. For more complex scenarios (such as search) you can also defined specific behavior by passing false in the second argument.