fix(pdf) set initial pdf scale to fit to available content height
fix #805
fix #622

Background:
In 3.1.1 we gained possibility to zoom in and out pdf :fireworks:
https://github.com/UniversalViewer/universalviewer/commit/888c4dfb231db69b2299602c5dce762994c1210e

The initial scale was set to 0.7 which in some cases made pdf smaller
than available content area, in other pdf was bigger.

To address this issue pull #742 was made, in which the canvas was
resized to available content height. Pdf was still rendered at 0.7 scale
resulting in blurry text.

This is another approach, which set the scale given the pdf page height
and available content size prior to rendering.